% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ap.R
\name{count_ap}
\alias{count_ap}
\title{Count associated positives}
\usage{
count_ap(x)
}
\arguments{
\item{x}{a list of scan_CommonRare results}
}
\value{
a vector of counts
}
\description{
Summarizes results from scan_CommonRare.
Counts number of times a non-focal p-value was lower or equal to the p-value from the focal region.
}
\author{
Marcin Kierczak
}
