% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_by_maf.R
\name{count_by_maf}
\alias{count_by_maf}
\title{count number of loci below and above a maf threshold}
\usage{
count_by_maf(regions, vcf_file, maf_threshold = 0.01)
}
\arguments{
\item{regions}{list of regions as returned by the read_regions_bed function}

\item{vcf_file}{path to vcf file containing genotypes}

\item{maf_threshold}{maf threshold for counting loci}
}
\value{
a tibble with region name, total count of loci, count of loci above and below the maf threshold
}
\description{
given region coordinates, the function extracts genotypes for all loci in the region,
computes minor allele frequency for each marker and returns the number of loci with maf <= threshold and
the number of loci maf > threshold.
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
