% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_allele_encoding.R
\name{fix_allele_encoding}
\alias{fix_allele_encoding}
\title{Fixes allele encoding so that the minor allele homozygote is always encoded as 2}
\usage{
fix_allele_encoding(G)
}
\arguments{
\item{G}{- goenotype matrix}
}
\value{
genotype matrix G (row - ind, col - marker) with fixed encoding:
0 - major allele homozygote,
1 - heterozygote,
2 - minor allele homozygote
being simply the count of the minor allele
}
\description{
Fixes allele encoding so that the minor allele homozygote is always encoded as 2
}
\details{
The genotype matrix G should be returning counts of minor allele.
If the major allele is not the reference allele, one needs to fix encoding and this
function takes care of this task. It counts reference and alternative allele and
if the reference allele is the minor allele, it swaps encoding for homozygotes.
}
\author{
Marcin Kierczak <marcin.kierczak__REPLACE_WITH_AT__scilifelab.se>
}
