% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_effects.R
\name{get_effects}
\alias{get_effects}
\title{Get weights for markers using beta distribution.}
\usage{
get_effects(maf, N, shape12, thr = 0.01, rare = T, frac_negative = 0, seed = F)
}
\arguments{
\item{maf}{- vector of minor-allele frequencies}

\item{N}{- number of markers to be assigned an effect}

\item{shape12}{- a two element vector of Beta distribution shapes.}

\item{rare}{- a boolean, if TRUE, markers below and equal to the `thr` will be sampled, i.e. the rare variants}

\item{frac_negative}{- fraction of effects to be set to negative}

\item{seed}{- set seed for sampling, if FALSE, default `sample` seed will be used}
}
\value{
a list with two elements: `marker_idx` is a vector of indices of the markers assigned an effect and `effects` is a vector of
effects, one for each marker
}
\description{
Given a vector of minor-allele frequencies (maf) for a number of markers and a number of markers to assign an effect to,
the function randomly selects N markers below or above provided maf threshold (default 1%) and assigns effect of a size sampled
from Beta distribution with given parameters so that the magnitude of the effect depends on the allele frequency. Desired fraction
of the effects will be negative. Monomorphic markers with `maf = 0` will be automatically excluded from sampling!
}
\author{
Marcin Kierczak <marcin.kierczak_NO_SPAM_scilifelab.se>
}
