% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotypes_seqminer.R
\name{get_genotypes_seqminer_matrix}
\alias{get_genotypes_seqminer_matrix}
\title{Get genotypes matrix (counts of reference allele) from one of the seqminer
read functions that return genotype matrix}
\usage{
get_genotypes_seqminer_matrix(x, ind_names = NULL, marker_names = NULL)
}
\arguments{
\item{x}{an object returned by seqminer *Matrix* function}

\item{ind_names}{names of the individuals to extract}

\item{marker_names}{names of the markers to extract}
}
\value{
genotype matrix ind x marker with counts of the reference allele
}
\description{
Get genotypes matrix (counts of reference allele) from one of the seqminer
read functions that return genotype matrix
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
