% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotypes_vcfR.R
\name{get_genotypes_vcfR}
\alias{get_genotypes_vcfR}
\title{Get genotypes matrix (counts of reference allele) from a vcfR object}
\usage{
get_genotypes_vcfR(x, marker_names)
}
\arguments{
\item{x}{a vcfR object}

\item{marker_names}{names of markers to extract}
}
\value{
genotype matrix ind x marker with counts of the reference allele
}
\description{
Get genotypes matrix (counts of reference allele) from a vcfR object
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
