% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maf.R
\name{get_maf}
\alias{get_maf}
\title{Get minor allele frequency vector}
\usage{
get_maf(x)
}
\arguments{
\item{x}{a gwaa-data or a vcfR object}
}
\value{
vector of minor allele frequencies
}
\description{
Get minor allele frequency vector
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
