% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hush.R
\name{hush}
\alias{hush}
\title{Silence too verbose code}
\usage{
hush(code)
}
\arguments{
\item{code}{the code to be silenced}
}
\value{
silenced result of code execution
}
\description{
Silence too verbose code
}
\author{
Marcin Kierczak <marcin.kierczak__INSERT_AT__scilifelab.se>
}
