% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_G.R
\name{impute_G}
\alias{impute_G}
\title{Impute missing genotypes using minor allele frequencies}
\usage{
impute_G(G, maf)
}
\arguments{
\item{G}{- genotype matrix (individuals in columns, markers in rows) to impute}

\item{maf}{- a vector with minor allele frequencies for the population from which G comes}
}
\value{
an imputed matrix
}
\description{
Impute missing genotypes using minor allele frequencies
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
