% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Plots effect sizes distribution dependent on minor allele frequency.}
\usage{
plot_effects(shapes_rare, shapes_common, thr)
}
\arguments{
\item{shapes_rare}{- params for Beta distribution for the rare variants}

\item{shapes_common}{- params for Beta distribution for the common variants}

\item{thr}{- threshold, markers with maf below or equal to it are considered rare}
}
\description{
Plots effect sizes distribution dependent on minor allele frequency.
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
