% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrk_density.R
\name{plot_mrk_density}
\alias{plot_mrk_density}
\title{Plot density of markers in a region}
\usage{
plot_mrk_density(coords, w = 1000, plot = F)
}
\arguments{
\item{coords}{- coordinates of all markers}

\item{w}{- window (interval) size, e.g. 1000 for 1kb}

\item{plot}{- a logical indicating whether to plot density}
}
\value{
a vector of densities
}
\description{
Plot density of markers in non-overlaping intervals of given size
}
\author{
Marcin Kierczak
}
