% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_region_vcf.R
\name{read_region_vcf}
\alias{read_region_vcf}
\title{read specific region from VCF file to a gwasim type genotype matrix}
\usage{
read_region_vcf(locus, vcf_file, force_silent = F, GP_map = c(0, 1, 2))
}
\arguments{
\item{locus}{region to extract, e.g. 22:17638600-17641201}

\item{vcf_file}{path to vcf file containing genotypes}

\item{force_silent}{silence seqminer verbosity (may cause some issues on non-*nix machines)}

\item{GP_map}{a genotype-phenotype map, by default AA-0, Aa-1 and aa-2}
}
\value{
a list with genotypes matrix and maf vector or NULL if region cannot be extracted
}
\description{
given region coordinates (locus), the function extracts genotypes for all loci in that region,
constructs genotypes matrix G (applies imputation and fixes allele encoding) and computes maf per marker.
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
