% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_regions_bed.R
\name{read_regions_bed}
\alias{read_regions_bed}
\title{Read regions information from bed file.}
\usage{
read_regions_bed(bed_file)
}
\arguments{
\item{bed_file}{path to the standard headerless bed file}
}
\value{
a tibble with region name,chromosome, start coord, end coord (in VCF 1-based coords) and region size
}
\description{
Read in regions from a bed file (no header allowed).
}
\details{
Only the first three columns will be read in.
chromosome names should begin with 'chr'. The returned coordinates are in the VCF coordinates (1-based).
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
