% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_G.R
\name{recode_G}
\alias{recode_G}
\title{Re-code genotypes according to the supplied genotype-phenotype map}
\usage{
recode_G(G, GP_map)
}
\arguments{
\item{G}{genotypes matrix}

\item{GP_map}{genotype-phenotype map}
}
\value{
weights matrix
}
\description{
Given a genotype-phenotype (GP) map recodes genotypes in G.
The GP map is a matrix with each row containing new coding for one marker in G:
new codes for genotypes 'AA', 'aA' and 'aa' (in G encoded as 0, 1 and 2, respectively).
}
\details{
If only a vector GP map is provided, same re-coding is assumed for each and every
locus in G.
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
