% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm_betas.R
\name{rnorm_betas}
\alias{rnorm_betas}
\title{Effect sizes $\Beta$ from a normal distribution.}
\usage{
rnorm_betas(maf, mean = 0, sd = 1)
}
\arguments{
\item{maf}{vector of minor-allele frequencies.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}
}
\value{
vector of effect sizes
}
\description{
Given a vector of minor-allele frequencies (maf), mean and standard deviation, sample `length(maf)`
effects from a normal distribution.
}
\author{
Marcin Kierczak <marcin.kierczak_NO_SPAM_scilifelab.se>
}
