% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_CommonRare.R
\name{scan_CommonRare}
\alias{scan_CommonRare}
\title{Run SKAT-O CommonRare model on all regions}
\usage{
scan_CommonRare(focal_region, regions, null_model, vcf_file)
}
\arguments{
\item{focal_region}{- name of the focal region}

\item{regions}{- list of regions to scan}

\item{null_model}{- SKAT null model}

\item{vcf_file}{- path to vcf file with genotypes}
}
\value{
a tibble with association p-value for every region and info whether
the region is the focal region
}
\description{
Given a phenotype simulated for a focal region, scan for
associations in all supplied regions.
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
