% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotype.R,
%   R/simulate_phenotype_at_locus.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Simulate phenotype given a set of common and rare variants.}
\usage{
simulate_phenotype(x, locus, means = c(1, 2, 3), std_devs = c(1, 1, 1))

simulate_phenotype(x, locus, means = c(1, 2, 3), std_devs = c(1, 1, 1))
}
\arguments{
\item{x}{a vcf object from vcfR package}

\item{thr_common_rare}{maf threshold for common and rare alleles}

\item{N_rare}{number of rare alleles to affect the trait}

\item{beta_params_rare}{parameters of the Beta distribution to simulate effect of rare alleles}

\item{perc_negative_rare}{how many (per cent) rare alleles have negative effect}

\item{N_common}{number of common alleles to affect the trait}

\item{beta_params_common}{parameters of the Beta distribution to simulate effect of common alleles}

\item{perc_negative_common}{how many (per cent) common alleles have negative effect}

\item{e}{a vector of mean and sd for the error term. Default (0, 1).}
}
\value{
a vector of simulated phenotypes

a vector of simulated phenotypes
}
\description{
Simulate phenotype given a set of common and rare variants.

Simulate phenotype for a single locus based on 3 sets of dieffect distribution parameters.
}
\author{
Marcin Kierczak <marcin.kierczak__INSERT_AT__scilifelab.se>

Marcin Kierczak <marcin.kierczak__INSERT_AT__scilifelab.se>
}
