% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotypes.R
\name{get_genotypes}
\alias{get_genotypes}
\title{Get genotypes matrix (counts of reference allele) from a GenABEL gwaa-data, vcfR object or
an object imported using seqminer}
\usage{
get_genotypes(x, ind_names = NULL, marker_names = NULL)
}
\arguments{
\item{x}{a gwaa-data, a vcfR object or a genotype matrix loaded using seqminer}

\item{ind_names}{names of individuals to extract}

\item{marker_names}{names of markers to extract}
}
\value{
genotype matrix ind x marker with counts of the reference allele
}
\description{
Get genotypes matrix (counts of reference allele) from a GenABEL gwaa-data, vcfR object or
an object imported using seqminer
}
\details{
if marker_names is not provided, all markers present in x will be returned
similarly, if ind_names is not present all individuals will be returned
NOTE: beware that the genotype matrix returned contains counts of a reference allele
which NOT NECESSARILY is the minor allele. Use gwasim::validate_maf() to be sure
you are working with the minor alllele counts. Also, note that removing individuals
may change allele frequencies and thus different alllele will be considered minor.
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
