% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genotypes_gwaa_data.R
\name{get_genotypes_gwaa_data}
\alias{get_genotypes_gwaa_data}
\title{Get genotypes matrix (counts of reference allele) from a GenABEL gwaa-data object}
\usage{
get_genotypes_gwaa_data(x, ind_names = NULL, marker_names = NULL)
}
\arguments{
\item{x}{a gwaa.data object (GenABEL internal format)}

\item{ind_names}{names of the individuals to extract}

\item{marker_names}{names of the markers to extract}
}
\value{
genotype matrix ind x marker with counts of the reference allele
}
\description{
Get genotypes matrix (counts of reference allele) from a GenABEL gwaa-data object
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
