% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_G.R
\name{impute_G}
\alias{impute_G}
\title{Impute missing genotypes using minor allele frequency}
\usage{
impute_G(G)
}
\arguments{
\item{G}{- genotype matrix (individuals in columns, markers in rows) to impute}
}
\value{
an imputed matrix
}
\description{
Imputes missing genotypes using minor allele frequency and assuming Hardy-Weinberg equilibrium
}
\author{
Marcin Kierczak <marcin.kierczak@scilifelab.se>
}
