% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_y_in_regions.R
\name{sim_y_in_regions}
\alias{sim_y_in_regions}
\title{Simulate phenotypes in a number of regions}
\usage{
sim_y_in_regions(
  regions,
  vcf_file,
  n_markers,
  get_betas_fun = rnorm_betas,
  get_betas_args = list(mean = 0, sd = 1),
  rare = T,
  frac_negative = 0,
  thr = 0.01,
  force_silent = T,
  seed = 42,
  e_mean = 0,
  e_sd = 1,
  GP_map = c(0, 1, 2)
)
}
\arguments{
\item{regions}{list of regions to use.}

\item{vcf_file}{character, path to vcf file containing genotypes.}

\item{n_markers}{integer, number of markers to use in the simulation (see \link{get_effects}.}

\item{get_betas_fun}{see \link{get_effects}.}

\item{get_betas_args}{see \link{get_effects}.}

\item{rare}{see \link{get_effects}.}

\item{frac_negative}{see \link{get_effects}.}

\item{thr}{see \link{get_effects}.}

\item{force_silent}{logical, if `TRUE` seqminer verbosity will be silenced (may cause some issues on non-*nix machines).}

\item{seed}{see \link{get_effects}.}

\item{e_mean}{numeric, error distribution mean.}

\item{e_sd}{numeric, error distribution standard deviation.}

\item{GP_map}{vector of genotype-phenotype map values for `AA`, `aA` and `aa` respectively.}
}
\value{
list with character region name and a vector of simulated phenotypes.
}
\description{
given region coordinates (locus), the function extracts genotypes for all loci in that region,
constructs genotypes matrix G (applies imputation and fixes allele encoding) and computes maf per marker.
}
\author{
Marcin Kierczak \email{marcin.kierczak@scilifelab.se}
}
