% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_phenotype.R
\name{simulate_phenotype}
\alias{simulate_phenotype}
\title{Simulate phenotype given a set of common and rare variants.}
\usage{
simulate_phenotype(
  x,
  thr_common_rare = 0.01,
  N_rare,
  beta_params_rare,
  perc_negative_rare,
  N_common,
  beta_params_common,
  perc_negative_common,
  e = c(0, 1)
)
}
\arguments{
\item{x}{a vcf object from vcfR package}

\item{thr_common_rare}{maf threshold for common and rare alleles}

\item{N_rare}{number of rare alleles to affect the trait}

\item{beta_params_rare}{parameters of the Beta distribution to simulate effect of rare alleles}

\item{perc_negative_rare}{how many (per cent) rare alleles have negative effect}

\item{N_common}{number of common alleles to affect the trait}

\item{beta_params_common}{parameters of the Beta distribution to simulate effect of common alleles}

\item{perc_negative_common}{how many (per cent) common alleles have negative effect}

\item{e}{a vector of mean and sd for the error term. Default (0, 1).}
}
\description{
Simulate phenotype given a set of common and rare variants.
}
\author{
Marcin Kierczak <marcin.kierczak__INSERT_AT__scilifelab.se>
}
